Implement offer logic
The offer logic allows for any number of offers for each item, but currently multiple offers can never be applied to the same subset of items. For example, if you have an offer for 6 items, an offer for 3, and another offer for 1, the largest offer will consume as many as possible, then on down until all quantities are consumed.